+++ module/Config/src/Config/Controller/RoleController.php
@@ -89,11 +89,11 @@ class RoleController extends Action
                 $role_model->addData($form->getInputFilter()->getValues());
                 $role_model->save();
 
-                $this->flashMessenger()->setNamespace('success')->addMessage('Role saved!');
+                $this->flashMessenger()->addSuccessMessage('Role saved!');
                 return $this->redirect()->toRoute('userRoleEdit', array('id' => $role_id));
             }
 
-            $this->flashMessenger()->setNamespace('error')->addMessage('Role can not saved!');
+            $this->flashMessenger()->addErrorMessage('Role can not saved!');
             $this->useFlashMessenger();
         }
 
@@ -112,11 +112,11 @@ class RoleController extends Action
         {
             if($role_model->delete())
             {
-                return $this->returnJson(array('success' => TRUE, 'message' => 'Role deleted!'));
+                return $this->returnJson(array('success' => TRUE, 'message' => 'Role has been deleted'));
             }
         }
 
-        return $this->returnJson(array('success' => FALSE, 'message' => 'Role does not exists!'));
+        return $this->returnJson(array('success' => FALSE, 'message' => 'Role does not exists'));
     }
 
     /**
@@ -131,7 +131,7 @@ class RoleController extends Action
         $role_model = Role\Model::fromId($role_id);
         if($role_model->getName() === Role\Model::PROTECTED_NAME)
         {
-            $this->flashMessenger()->setNamespace('error')->addMessage("Can't edit this role");
+            $this->flashMessenger()->addErrorMessage("Can't edit this role");
             return $this->redirect()->toRoute('userRole');
         }
 
@@ -148,11 +148,11 @@ class RoleController extends Action
                 $role_model->addData($form->getInputFilter()->getValues());
                 $role_model->save();
 
-                $this->flashMessenger()->setNamespace('success')->addMessage('Role saved!');
+                $this->flashMessenger()->addSuccessMessage('Role saved!');
                 return $this->redirect()->toRoute('userRoleEdit', array('id' => $role_id));
             }
 
-            $this->flashMessenger()->setNamespace('error')->addMessage('Role can not saved!');
+            $this->flashMessenger()->addErrorMessage('Role can not saved!');
             $this->useFlashMessenger();
         }